+++ orix/base/__init__.py
@@ -18,6 +18,8 @@
 
 import numpy as np
 
+from orix._util import deprecated
+
 
 # Lists what will be imported when calling "from orix.base import *"
 __all__ = [
@@ -110,10 +112,20 @@ class Object3d:
         return self.data.shape[:-1]
 
     @property
+    @deprecated(since="0.8", alternative="ndim", removal="0.9", object_type="property")
     def data_dim(self):
         """int : The dimensions of the data.
 
-        For example, if `data` has shape (4, 4, 3), `data_dim` is 3.
+        For example, if `data` has shape (4, 5, 6), `data_dim` is 3.
+
+        """
+        return self.ndim
+
+    @property
+    def ndim(self):
+        """int : The number of navigation dimensions of the instance.
+
+        For example, if `data` has shape (4, 5, 6), `data_dim` is 3.
 
         """
         return len(self.shape)